better handling of chess pieces
