+++ chess_pieces/bishop.py
@@ -1,4 +1,4 @@
-from chess_pieces import AbstractChessPiece
+from chess_pieces.abstract import AbstractChessPiece
 
 
 class Bishop(AbstractChessPiece):
