+++ chess_pieces/king.py
@@ -1,4 +1,4 @@
-from chess_pieces import AbstractChessPiece
+from chess_pieces.abstract import AbstractChessPiece
 
 
 class King(AbstractChessPiece):
