+++ chess_pieces/knight.py
@@ -1,5 +1,4 @@
-from chess_pieces import AbstractChessPiece
-
+from chess_pieces.abstract import AbstractChessPiece
 
 class Knight(AbstractChessPiece):
