+++ chess_pieces/rook.py
@@ -1,5 +1,4 @@
-from chess_pieces import AbstractChessPiece
-
+from chess_pieces.abstract import AbstractChessPiece
 
 class Rook(AbstractChessPiece):
