+++ tests/board/test_players.py
@@ -1,5 +1,6 @@
 from board.players import Player, ChessPlayers
-from chess_pieces import Color
+from chess_pieces import Color, Group
+from chess_pieces.pawn import Pawn
 
 
 def test_player_to_player_comparison():
@@ -22,6 +23,22 @@ def test_player_to_string_comparison():
     player = Player(Color.white)
     assert player == 'white'
     assert player != 'black'
+
+
+def test_player_add_chess_piece():
+    """
+    Test that a chess piece can be added
+    to Player instance.
+    """
+    player_1 = Player(Color.white)
+    pawn = Pawn(
+        position=(0, 0),
+        piece_nr=1,
+        group=Group.lower,
+        color=Color.white
+    )
+    player_1.add_chess_piece(pawn)
+    assert len(player_1.chess_pieces['Pawn']) == 1
 
 
 def test_chess_players():
