+++ tests/pieces/test_chess_pieces.py
@@ -1,4 +1,4 @@
-from chess_pieces import AbstractChessPiece
+from chess_pieces.abstract import AbstractChessPiece
 from chess_pieces.schema import Color, Group
